% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTheBenchmarkData.R
\name{download_benchmark_data}
\alias{download_benchmark_data}
\title{download benchmark files}
\usage{
download_benchmark_data(xx, outdir = ".", simulate = FALSE)
}
\arguments{
\item{xx}{list of list of files as generated by list_benchmark_data}

\item{outdir}{destination folder}

\item{simulate}{default FALSE (download data)}
}
\description{
download benchmark files
}
