% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTheBenchmarkData.R
\name{downloadBenchmarks}
\alias{downloadBenchmarks}
\title{download benchmark datasets to outdir which must exist}
\usage{
downloadBenchmarks(
  url = "http://fgcz-ms.uzh.ch/~wolski/protriple_DIA_data/",
  outdir = ".",
  simulate = FALSE
)
}
\arguments{
\item{url}{data to benchmark http://fgcz-ms.uzh.ch/~wolski/protriple_DIA_data/"}

\item{outdir}{destination folder}

\item{simulate}{default FALSE (download data)}
}
\description{
download benchmark datasets to outdir which must exist
}
