% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTheBenchmarkData.R
\name{list_benchmark_data}
\alias{list_benchmark_data}
\title{download benchmark datasets to outdir which must exist}
\usage{
list_benchmark_data(url = "http://fgcz-ms.uzh.ch/~wolski/protriple_DIA_data/")
}
\arguments{
\item{url}{data to benchmark http://fgcz-ms.uzh.ch/~wolski/protriple_DIA_data/"}

\item{outdir}{destination folder}

\item{simulate}{default FALSE (download data)}
}
\description{
download benchmark datasets to outdir which must exist
}
