% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specnaut.R
\name{sanitize_specnaut}
\alias{sanitize_specnaut}
\title{massage spectronaut table}
\usage{
sanitize_specnaut(prec)
}
\description{
select columns, count peptides (stripped sequences per protein), summarize FG.Quantity
}
